Refactor : 텍스트에디터 구현을 lexical 폴더로 분리
