+++ src/lexical/components/Buttons/codeButton.tsx
@@ -1,9 +1,9 @@
-import { codeNode } from '#/components/plugins/blockTypes'
-import BaseToolButton from '#/components/plugins/Buttons/baseToolButton'
-import { buttonSizes, CommonToolButtonProps } from '#/components/plugins/Buttons/buttonTypes'
-import ToolbarIcon from '#/components/plugins/Buttons/toolbarIcon'
-import { CODE_LANGUAGE_COMMAND } from '#/components/plugins/codeHighlightPlugin'
-import { formatCode } from '#/components/plugins/utils'
+import { codeNode } from '#/lexical/plugins/blockTypes'
+import BaseToolButton from '#/lexical/components/Buttons/baseToolButton'
+import { buttonSizes, CommonToolButtonProps } from '#/lexical/components/Buttons/buttonTypes'
+import ToolbarIcon from '#/lexical/components/Buttons/toolbarIcon'
+import { CODE_LANGUAGE_COMMAND } from '#/lexical/plugins/codeHighlightPlugin'
+import { formatCode } from '#/lexical/plugins/utils'
 import { cls } from '#/libs/client/utils'
 import { CODE_LANGUAGE_FRIENDLY_NAME_MAP } from '@lexical/code'
 import { $getSelection, $isRangeSelection } from 'lexical'
