+++ src/lexical/editor.tsx
@@ -7,17 +7,17 @@ import { ContentEditable } from '@lexical/react/LexicalContentEditable'
 import { HistoryPlugin } from '@lexical/react/LexicalHistoryPlugin'
 import { LexicalErrorBoundary } from '@lexical/react/LexicalErrorBoundary'
 import { ComponentProps, ReactNode, useState } from 'react'
-import { nodes } from '#/nodes/EditorNodes'
-import { ToolbarPlugin } from '#/components/plugins/toolbarPlugin'
-import { theme } from '#/components/editorTheme'
+import { nodes } from '#/lexical/nodes/EditorNodes'
+import { ToolbarPlugin } from '#/lexical/plugins/toolbarPlugin'
+import theme from '#/lexical/styles/editorTheme'
 import { ListPlugin } from '@lexical/react/LexicalListPlugin'
 import { CheckListPlugin } from '@lexical/react/LexicalCheckListPlugin'
-import CodeHighlightPlugin from '#/components/plugins/codeHighlightPlugin'
+import CodeHighlightPlugin from '#/lexical/plugins/codeHighlightPlugin'
 import { TextInput } from '#/components/Inputs'
 import { TabIndentationPlugin } from '@lexical/react/LexicalTabIndentationPlugin'
-import InlineToolbarPlugin from '#/components/plugins/inlineToolbarPlugin'
+import InlineToolbarPlugin from '#/lexical/plugins/inlineToolbarPlugin'
 import { LinkPlugin } from '@lexical/react/LexicalLinkPlugin'
-import FloatingLinkEditorPlugin from '#/components/plugins/FloatingLinkEditorPlugin'
+import FloatingLinkEditorPlugin from '#/lexical/plugins/FloatingLinkEditorPlugin'
 
 function onError(error: unknown) {
   console.error(error)
