+++ src/lexical/nodes/EditorNodes.ts
@@ -3,6 +3,7 @@ import { ListItemNode, ListNode } from '@lexical/list'
 import { Klass, LexicalNode } from 'lexical'
 import { CodeNode, CodeHighlightNode } from '@lexical/code'
 import { LinkNode } from '@lexical/link'
+import { StickyNode } from '#/lexical/nodes/StickyNode'
 
 //
 // Klass<T> : 제네릭 T 타입의 클래스의 인스턴스를 생성할 수 있는 클래스만 허용하는 유틸리티 타입
@@ -15,4 +16,5 @@ export const nodes: Klass<LexicalNode>[] = [
   CodeNode,
   CodeHighlightNode,
   LinkNode,
+  StickyNode,
 ]
