+++ src/lexical/plugins/codeHighlightPlugin.tsx
