+++ src/lexical/styles/editorTheme.module.css.d.ts
